requirements.txt
openai
litellm
requests
beautifulsoup4
python-dotenv
tiktoken
rich
google-generativeai
google-cloud-aiplatform